OTTER-556: converge file preview modal across DO and RL flows
Move CodeViewer into components/file-viewers and add a `codeViewer`
match function to the FileViewer registry so any path with a known code
extension is syntax-highlighted automatically.

Extract FilePreviewModal into its own module (components/file-preview-modal)
that renders the shared FileViewer chain and an inline Download link in
its header. The DO encrypted-files panel now reuses this modal in place
of its local FileViewerModal; the per-row Download stays for the DO
table, and the same Download is also available from the modal header,
matching the RL submitted-code flow.
