+++ src/app/[orgSlug]/admin/settings/code-envs.tsx
@@ -39,8 +39,7 @@ import { Refresher } from '@/components/refresher'
 import { ErrorPanel } from '@/components/panel'
 import { LoadingMessage } from '@/components/loading'
 import { ActionSuccessType, DATA_SOURCE_TYPES, type DataSourceType } from '@/lib/types'
-import { CodeViewer } from '@/components/code-viewer'
-import { FileViewer } from '@/components/file-viewers'
+import { CodeViewer, FileViewer } from '@/components/file-viewers'
 import { useState } from 'react'
 import { isActionError } from '@/lib/errors'
 import type { OrgCodeEnvSettings, ScanStatus } from '@/database/types'
